refactor: complete Book domain DDD migration
- Moved BookRepository to domain and implemented SqliteBookRepository in infrastructure
