--- src-tauri/src/database/book/mod.rs
@@ -1,7 +1,0 @@
-pub mod model;
-pub mod repository;
-pub mod sqlite_repository;
-
-pub use model::*;
-pub use repository::*;
-pub use sqlite_repository::*;
