+++ src-tauri/src/database/mod.rs
@@ -1,4 +1,3 @@
-pub mod book;
 pub mod bookshelf;
 pub mod series;
 pub mod tag;
