+++ src-tauri/src/database/tag/repository.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 
-use crate::database::book::BookWithState;
+use crate::domain::book::entity::BookWithState;
 
 use super::model::Tag;
